AA-80 경매 시스템 구현

+++ src/main/java/com/imagine/another_arts/domain/art/service/dto/response/BasicGeneralArtResponse.java
@@ -14,7 +14,7 @@ public class BasicGeneralArtResponse {
     private BigInteger artId; // 작품 ID
     private String artName; // 작품명
     private String artDescription; // 작품 설명
-    private Integer artInitPrice; // 작품 초기 가격
+    private Long artInitPrice; // 작품 초기 가격
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     private Timestamp artRegisterDate; // 작품 등록 날짜
     private String artStorageName; // 작품 서버 저장명 (UUID)
